UC-782 feat: new scene for the violence ending
The audio files, project files, and video files referents to the violence ending ware added to the repository.

A new scene for the violence ending was created.

When started, the scene auto play the properly animation and audio for the violence ending.

After the video is finished, the scene automatically goes to the Main Title scene.